Update deploy_p0f.sh
I couldn't sleep with my commit, as it doesn't catch any other scenarios, so i modified it a little. It catches too many interfaces or non now, as well as you are able to enter them manually in general or if you run into a scenario mentioned before. Just tested on an ubuntu 16.04 vm.